Added - CBA Settings & Log function & Beginning of stringtable
What
=================
Added: CBA Settings to allow the settings normally set by the 3DEN
module to be set by CBA Settings as well/instead.
Added: Log function so that we can see what's happening when we're
debugging.
Added: Beginning of a stringtable, to allow for translations to be
implemented by any others who'd care to contribute.

Issue ID: N/A